ARROW-15707: [C++][FlightRPC] Make Flight data plane tests reusable
