+++ cpp/src/arrow/flight/CMakeLists.txt
@@ -213,6 +213,7 @@ if(ARROW_TESTING)
                 OUTPUTS
                 ARROW_FLIGHT_TESTING_LIBRARIES
                 SOURCES
+                test_definitions.cc
                 test_util.cc
                 DEPENDENCIES
                 GTest::gtest
